The only Introduction sentence omits half the description.
Method Enum.IsDefined is useful when the program has a integer value to test, and the lone sentence that comprises the entire Introduction mentions this.  But .IsDefined is equally useful when instead the program has a string that might match the Name that is assigned to the integer (integral value) inside the enum body; yet the Introduction improperly omits this important fact and functionality.
I realize this text change really needs to be made in the .NET source code ///summary for the System.Enum class, for the text change to live on indefinitely.