Correct TreeViewItem.IsSelectionActive description
The docs incorrectly stated that TreeViewItem.IsSelectionActive indicated whether the TreeViewItem itself has the keyboard focus. By inspection this is not true. In fact, if the containing TreeView has the keyboard focus then *every* TreeViewItem in it will return true from IsSelectionActive. So the docs need to make it clear that although this property is defined as a member of TreeViewItem, it actually reports information about the containing TreeView.

This can be demonstrated by experiment. Moreover, it appears to be by design. The TreeViewItem.IsSelectionActive property is in fact an alias for Selector.IsSelectionActive, an inherited property which is already correctly documented as reporting whether the containing Selector has the focus.

Furthermore, inspecting the reference source reveals not only that this property is an alias, but also that WPF makes no attempt to make TreeViewItem.IsSelectionActive specific to any single TreeViewItem. The handling for this property exists entirely at the TreeView level.