Made clear DNS subsystem is used
Documentation was confusing because is worded as if it queried using the DNS protocol, not the subsystem. If it was just using the DNS protocol, it could be bypassing the hosts file, similar to the nslookup tool. Confirmed this code does not bypass the hosts file. See issue: https://github.com/dotnet/docs/issues/8418#issuecomment-432017794