M32642: Truncated source
Hello @mairaw,  @dend, @rpetrusha,
Translator has reported possible source content issue. 
Description: Would you please kindly advise how to translate these truncated sources - whose live pages are not found on the web?
Full Source String:
When the method returns, the manager of
When the method returns, the offset from the start of the
When the method returns, the length of the
When the method returns, an instance of
that the
represents.

Please review the comment tag added and reply with explanation if fix is needed or not. If you make related fix in another PR then share your PR number so we can confirm and close this PR.
Many thanks in advance.