Different cross platform behavior of GetFileName
As it was discussed in this issue https://github.com/dotnet/coreclr/issues/21597 the behavior of GetFileName is not the same and cannot be the same on Windows and Linux because `'\\'` is a legal file name on Linux. 

It was suggested on that issue discussion that this could be better clarified on the documentation, hence this proposed change.