Update MailAddress.xml
Sorry for the delay. Between figuring out the tool and the subtly of the language I have been slow. I think I now see why I did not understand this the first time through I have provided something that should be more clear. Instead of the example output I have specifically referred to the standard format as it now appears to me that the whole point of these sentences is to highlight that two non-standard forms can be automatically converted to the standard form.