Specify how to identify culture
The docs refer to [MS-LCID](https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-lcid/a9eac961-e77d-41a6-90a5-ce1a8b0cdb9c) for a list of cultures, but don't say how you should refer to a culture. This change clarifies that the language tag (as specified in the 4th column of the table in MS-LCID) is used.