Corrected (?) F# signature of Receive method
The `Receive` method's F# signature was incorrect—the remoteEP parameter was missing:

``` xml
<MemberSignature Language="F#" Value="member this.Receive :  -&gt; byte[]" Usage="udpClient.Receive remoteEP" />
```

Changed the signature and the usage to the following:

``` xml
<MemberSignature Language="F#" Value="member this.Receive : (remoteEP : IPEndPoint ref) -&gt; byte[]" Usage="udpClient.Receive (ref remoteEP)" />
```

However, I am not quite sure how correct this is. The 'Value' attribute is probably correct, but the 'Usage' attribute is suspect. One would not simply pass in `ref remoteEP` but rather something more like this:

``` fsharp
let remoteEndPointRef = ref (IPEndPoint (IPAddress.Any, 0))
let bytes = udpClient.Receive remoteEndPointRef
printfn "Received from %A: %A" !remoteEndPointRef bytes
```

Finally, there is no F# example code. I should love to submit an F# version of the C# sample, but it is not at all clear to me how to add it!

The C# code sample is apparently an include of some sort:

```
[!code-csharp[Classic UdpClient.PublicMethodsAndPropertiesExample#11](~/samples/snippets/csharp/VS_Snippets_Remoting/Classic UdpClient.PublicMethodsAndPropertiesExample/CS/source.cs#11)]
```

Is there a way to do something similar to include an F# sample?