MouseWheel properties are equivalent
This was never clearly documented. We supported IntelliMouse drivers on Windows 9x systems. `NativeMouseWheelSupport` told you if the OS natively had a mouse wheel installed. `MouseWheelSupport` told you if the OS had a mouse wheel installed OR an IntelliMouse driver was installed to give said support. The driver sent mouse wheel messages on 9x systems, but the OS didn't natively understand what that was.

Also removing the OS comments as they can be confusing and we're long long past those days. :)