Fix http status code dead link
The link pointing to the Http Status Codes is dead and redirects back to microsoft.com so fixed it. Seems like it pointing to msdn2.microsoft.com which is dead so just used the msdn.microsoft.com subdomain instead.

Can't create a go.microsoft.com link so used the long link instead.