Space characters are encoded as '+'
Existing description implies that the specified character classes are the only exceptions to hexadecimal encoding.
I'm trying to match Javascript 'encodeURIComponent' behavior, and expected WebUtility.UrlEncode to provide this. 
However, as spaces are a further exception, this method produces results that are functionally identical to that produced by HttpUtility.UrlEncode.

Might also be useful to mention Uri.EscapeDataString and AntiXss.UrlEncode as well here, as these are now preferred options when looking to encode URIs in a transparent, reversible manner.