Add sbyte overload when using Ldc_I4_S
The `ldc.i4.s` opcode accepts values in the range `[-128,127]`, but the docs only showed the overload with the `byte` parameter as supported when using this opcode. This was confusing for users (including me), since negative values can't obviously be cast to `byte`.
The `sbyte` overload works too when using this opcode.