Case for no specific HRESULT
A lot of user-defined exception types don't do anything with HRESULT. The way the documentation was written, it make it seem like each such type had to come up with its own HRESULTs. This makes it clear that it's OK to let the HRESULT remain at 0 to indicate that you're not doing anything with HRESULTs.