Note interactive logon session behavior
I noticed that users in interactive sessions can always view counter data, and only non-interactive logon sessions need to be in the Performance Monitoring Users group in order to view counter data, but the documentation here made no mention of the interactive vs non-interactive distinction, so I have added that information.

I noticed that this exception for interactive sessions was introduced in Vista SP1. In Vista RTM, users in interactive sessions had to be in the Performance Monitoring Users group in order to view counter data. I believe that is why this documentation does not mention the interactive session exception, because this documentation was probably written when Vista RTM was released and not updated since.