Fix unsafe Random use in ReaderWriterLock samples (#4099)
The ReaderWriterLock samples were sharing a single Random instance
across threads. The various .Next methods on Random instances are not
safe to be called concurrently.

Now, each thread has its own Random instance that it uses. Some of the
worker methods needed to be adjusted to take a Random instance as input.

I considered using a thread-local Random instance. However, the sample
is already complex, so I opted to avoid introducing thread-local values,
which would need to be briefly explained.