Wrong Link
I haven't made any changes, but want to point out the following.

In the Constructor Overloads section, when the user clicks on both the below links

RNGCryptoServiceProvider(CspParameters)
RNGCryptoServiceProvider(String)

both links are navigating to the section RNGCryptoServiceProvider(String). I can find the urls generated in the browser are correct and separated, but both the following URLs pointing to the same section in the page. 

https://docs.microsoft.com/en-us/dotnet/api/system.security.cryptography.rngcryptoserviceprovider.-ctor?view=netcore-3.1#System_Security_Cryptography_RNGCryptoServiceProvider__ctor_System_Security_Cryptography_CspParameters_

https://docs.microsoft.com/en-us/dotnet/api/system.security.cryptography.rngcryptoserviceprovider.-ctor?view=netcore-3.1#System_Security_Cryptography_RNGCryptoServiceProvider__ctor_System_String_