Fix remarks for Insert overload which takes CacheDependency (#4483)
The Remarks for the overload of `Insert` which takes a `CacheDependency` appears to have been copied and pasted from the overload which does not take that parameter and therefore stated that the object added to the cache would be inserted with "no file or cache dependencies", which of course is incorrect. This edit removes that text.