CultureInfo.CurrentCulture default on Unix-likes
The last step of the "How a Thread's Culture Is Determined" section of
`CultureInfo.CurrentCulture` currently documents calling the Windows
`GetUserDefaultLocaleName` function.  On Unix-like systems [ICU]
`uloc_getDefault` is called.  The current implementation of
`uloc_getDefault` eventually calls POSIX `setlocale(3)` with category
`LC_MESSAGES` to get the default, as described in #4254.

This commit documents the use of `uloc_getDefault` and its use of
`setlocale` (which I expect to be more familiar and searchable) on
Unix-like systems to help developers understand the default on these
systems.

Fixes: #4254

[ICU]: https://icu-project.org/

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>