Update FlagsAttribute.xml
The only effect of the `[Flags] attribute is to modify the behavior of calls to `ToString()`.  Although it is difficult to think of direct practical negative consequences of implying that `FlagsAttribute` is necessary to use an enumeration as a bit field in the way that the article does, a small unnecessary cognitive load is imposed when official documentation does not quite tally with one's experience.  Precision is worth it.