Compare CompareTo return with zero in examples
In the examples for `IComparable<T>.CompareTo`, the comparison operator
implementations are checking that `Double.CompareTo` returns exactly -1
or 1. The contract for `IComparable<T>.CompareTo` is that the return
value is negative, zero, or positive. A valid implementation of
`IComparable<T>.CompareTo` can return 87 or -2844587, for example.

Compare to zero instead of checking for equality with -1 and 1.

---

The managed C++ example returns true for the following code when it
should return false:

    Temperature^ zero = gcnew Temperature(0);
    return zero > zero;

Fix the C++ `operator>` implementation.