Update ApplicationException remarks
The current docs for `ApplicationException` state the it "Serves as the base class for application-defined exceptions" however the remarks section states "You should derive custom exceptions from the Exception class rather than the `ApplicationException` class." which are contradictory. I believe that the intention for the remarks statement is that libraries should not throw/catch `ApplicationException` since this type should only be used by applications directly.