Update DayOfWeek.xml
The ordering of the enum values is significant, both in terms of real-world calendar systems and the ordinality of the values. The previous document was not sorted by any obvious means, alphabetical might be acceptable at a stretch, but enums in C# have significant values and those values should are expected to define their _natural_ order.