Assembly.LoadFrom (.NET Framework) behaviour clarifications
The current docs can lead to confusion; they can be understood to say that an assembly with the same identity can only be loaded once - in fact it can only be loaded once in the same context, but the same identity can be both in the load and load-from contexts.

Also added a (significant, in my opinion) caveat with the influence of binding redirects.  Will remove if deemed too obscure.