Updated docs on CopyTo(Stream, Int32)
This method reads a specified number of bytes from the current stream, advances the read position and then writes them to the destination stream, advancing the write position of that as well.

The old documentation did not make it clear that the write or read positions were advanced at all. This can lead to a lot of confusion for programmers when they attempt to use this method.