Fix stream parameter description on GZipStream, BrotliStream, ZLibStream, DeflateStream
For the ctors that take a mode, the stream parameter is written to when CompressionMode.Compress is used or is read from when CompressionMode.Decompress is used.  These comments, in particular the ones on GZipStream, suggest the stream is written to regardless of the CompressionMode.  Also tweaked the ZLibStream comments to match.