Clarify Process.HasExited and the it raises Exited
The existing description implied that the Exited event could be raised "before" the user calls HasExited. This wasn't very clear in my opinion - so I've suggested a tweak to make it more explicit by revealing that the HasExited property accessor itself has a side-effect; invoking HasExited if the process has exited.