Remove wrong path case sensitive claims
#7989 describes that the Directory.Exists path parameter is case
sensitive on linux despite the documentation stating it is case
insensitive.

Presumably, this documentation comes from the Windows-only .NET
framework days where the file systems are expected to be case
insensitive.
This does not hold for newer multi-platform .NET versions which use
other case sensitive platforms.

As such, presumably, this holds for all Directory methods with a `path`
parameter. The statement is removed for every such method.

Fixes #7989