Missing event handler DownloadDataCompleted (#8551)
An example of subscribe event is provided for the WebClient events. Some events have the event handler method in the documentation (DownloadFileCompleted, DownloadProgressChanged, and UploadProgressChanged). Some events are missing the method specified when the event is subscribed. This includes DownloadDataCompleted which is missing the method to when the event was subscribed. DownloadDataCompleted was fixed to include the missing event handler method using the same pattern as UploadProgressChanged (a code section for the subscribe and a second code section for the event handler method)