Fx handler methods for event UploadProgressChanged
The documentation for WebClient.UploadProgressChanged event is as follows:
https://learn.microsoft.com/en-us/dotnet/api/system.net.webclient.uploadprogresschanged?view=net-6.0

This documentation for UploadProgressChanged uses <Snippet3>that originally included:
            client.UploadFileCompleted += new UploadFileCompletedEventHandler(UploadFileCallback2);
            client.UploadProgressChanged += new UploadProgressChangedEventHandler(UploadProgressCallback);

The documentation page for UploadProgressChanged then contains a reference to Snippet42 to document the above methods but Snippet42 contains the following methods:
UploadProgressCallback
DownloadProgressCallback: this is a download method not used in Snippet3

Snippet4 contains the documentation needed by Snippet3 -- the definition of the UploadFileCallback2 method. Snippet4 has no UploadProgressChangedEventHandler. I made a copy of the original UploadProgressCallback and called it UploadProgressCallback2. I placed this in Snippet4 and updated Snippet3 to reference UploadProgressCallback2 instead of UploadProgressCallback.

Once asyncmethods.cs is updated, I will change the source to this link:
https://learn.microsoft.com/en-us/dotnet/api/system.net.webclient.uploadprogresschanged?view=net-6.0

I will swap out the reference to Snippet42 and replace it with Snippet4.

I am pretty sure Snippet41 and Snippet42 are orphaned but this is obsolete documentation. I'd rather fix what is obviously broken (Snippet4) than purge unused code.