Update Instr(Int32, String, String)
We got a bug that InStr did not work because of confusion with this example. I could not work out adding a comment in the code, and this text change seemed more explanatory anyway.

(The example switches both the start position and the `CompareMethod` and the person who had a problem only noticed that the start position is changed.