Replaced 'T' type with 'TResult'
I think the statement on line 3059 

"Starting in .NET 6, for some `T` types and some result values, this method may return a cached singleton object rather than allocating a new object." 

is a bit confusing. I think it should be removed or enclosed in a box with a title like "Internal implementation".

Usually I/we as developer are not interested on the way a library is implemented. If someone wants to see the implementation details can browse the source code.