Update sample.cs
Minor cleanup:
  * In `Handle` setter, remove `else`.
  * In both the constructor and `Load` method, rename `Path` to `path`.
  * Deleted comment in setter explaining how if the handle is valid, return the handle value.
  * Got rid of brackets for if-statements with 1 line.
  * Changed constructor to be expression-bodied.
  * Deleted some unnecessary empty lines.

There's a good amount that could change for this example.

One more thing:
I came here to fix an error in the `Load` method as it appears in [MSDN](https://learn.microsoft.com/en-us/dotnet/api/microsoft.win32.safehandles.safefilehandle?view=net-7.0) (`&&` should be `||`). That is actually fixed here but the change hasn't been reflected in the site. Does something need to be triggered for this to happen?