Update NativeLibrary.xml
This changes the parameter **libraryPath** into **executableName**

Reasoning: 

The path is automatically handled by the abstraction of the API itself. The user is **not** supposed to provide a path: 

That would simply be not sensical, considering macOS, Linux, etc, all have different paths to begin with. 

The argument could never be multiple paths to all of them, and it's the executable name (that has to be on the PATH, obviously) that is our concern. 

I hope this is fine this way, and I also corrected two small typos. 
Thanks