Update README with API reference documentation details
The links from the api-docs repo has a bunch of general info about the doc updates and desired content style, but I didn't see anything that tells contibutors if their edit should be made in this repo or elsewhere. The runtime docs do talk about that, but those docs weren't easily discoverable from here.

I'm not sure if the front README is the right place for this, but I didn't see a better place.